backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0
pandas
numpy
scikit-learn
joblib
xgboost
azure-storage-blob